Refactor host routes into their own blueprint
