+++ natlas-server/app/host/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('host', __name__)
+
+from app.host import routes # noqa: F401
