+++ natlas-server/app/host/forms.py
@@ -1,0 +1,6 @@
+from flask_wtf import FlaskForm
+from wtforms import SubmitField
+
+
+class RescanForm(FlaskForm):
+	requestRescan = SubmitField('Request Rescan')
