+++ natlas-server/app/templates/errors/404.html
@@ -8,7 +8,7 @@
   </div>
   <div class="row">
   	<div class="col">
-		<p>Looks like we couldn't find what you're looking for. Maybe you'd prefer to go on <a href="{{url_for('main.randomHost')}}">an adventure</a>? Or maybe you'd like to <a href="{{ url_for('main.index') }}">return to browse</a>.</p>
+		<p>Looks like we couldn't find what you're looking for. Maybe you'd prefer to go on <a href="{{url_for('host.randomHost')}}">an adventure</a>? Or maybe you'd like to <a href="{{ url_for('main.index') }}">return to browse</a>.</p>
 	</div>
   </div>
 {% endblock %}
