+++ natlas-server/app/templates/host/versions/0.6.6/screenshots.html
@@ -14,7 +14,7 @@
 	{% for entry in historical_screenshots %}
 		<div class="row screenshot-header-row">
 			<h5 class="border-bottom">
-				<a href="{{url_for('main.host_historical_result', ip=ip, scan_id=entry.scan_id)}}">{{ entry.ctime | ctime }}</a>
+				<a href="{{url_for('host.host_historical_result', ip=ip, scan_id=entry.scan_id)}}">{{ entry.ctime | ctime }}</a>
 			</h5>
 		</div>
 		<div class="row image-row">
