+++ natlas-server/app/templates/includes/nav.html
@@ -14,7 +14,7 @@
               <a class="nav-link" href="{{ url_for('main.browseScreenshots') }}">Screenshots</a>
             </li>
             <li>
-              <a class="nav-link" href="{{ url_for('main.randomHost') }}">Random</a>
+              <a class="nav-link" href="{{ url_for('host.randomHost') }}">Random</a>
             </li>
           </ul>
